closes #4 fixed latitude and longitude data types

+++ src/main/java/com/openbanking/api/ng/controller/ATMController.java
@@ -40,8 +40,8 @@ public class ATMController {
     }
 
     @RequestMapping(value = "/latitude/{latitude}/longitude/{longitude}", method = RequestMethod.GET)
-    public List<ATM> getAtmsAtLocation(@PathVariable @ApiParam(value = "The location's latitude") int latitude,
-                                       @PathVariable @ApiParam(value = "The location's longitude") int longitude) {
+    public List<ATM> getAtmsAtLocation(@PathVariable @ApiParam(value = "The location's latitude") double latitude,
+                                       @PathVariable @ApiParam(value = "The location's longitude") double longitude) {
         return Collections.nCopies(3, new ATM());
     }
 }
