+++ src/main/java/com/openbanking/api/ng/payload/transaction/SingleTransfer.java
@@ -18,8 +18,8 @@ public class SingleTransfer {
     private String destinationNarration;
     private String transactionFee = "100";
     private String channel;
-    private int latitude = 100;
-    private int longitude = 101;
+    private double latitude = 100;
+    private double longitude = 101;
     private Date transactionDate;
     private PeriodType interval;
     private Date startDate;
@@ -83,19 +83,19 @@ public class SingleTransfer {
         this.channel = channel;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
